+++ BaseStation/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/BaseStation
+[install]
+install_scripts=$base/lib/BaseStation
